x86/cpu: Disable kernel LASS when patching kernel alternatives
Most of the kernel is mapped at virtual addresses
in the upper half of the address range. But kernel
deliberately initialized a temporary mm area
within the lower half of the address range
(a positive address range) for text poking, see
commit 4fc19708b165 ("x86/alternatives: Initialize
temporary mm for patching").

LASS stops access to a positive address (the lower
half address range) in kernel, and this can be
deactivated if AC bit in EFLAGS register is set.
Hence use stac and clac instructions around access
to the address to avoid triggering a LASS #GP fault.

Kernel objtool validation warns if the binary calls
to a non-whitelisted function that exists outside of
the stac/clac guard, or references any function with a
dynamic function pointer inside the guard; see section
9 in the document tools/objtool/Documentation/objtool.txt.

For these reasons, also considering text poking size is
usually small, simple modifications have been done
in function text_poke_memcpy() and text_poke_memset() to
avoid non-whitelisted function calls inside the stac/clac
guard.

Gcc may detect and replace the target with its built-in
functions. However, the replacement would break the
objtool validation criteria. Hence, add -fno-builtin
option to the CFLAGS for the file.

Signed-off-by: Yian Chen <yian.chen@intel.com>