sched/topology: Remove SD_PREFER_SIBLING flags from die domain on hybrid CPU
As mentioned by Ricardo, it is possible that SD_ASYM_PACKING and
SD_PREFER_SIBLING fight each other in the "DIE" domain on MTL:

Suppose there are 2 threads running on the compute die on an
otherwise idle system. When a CPU in the SoC die is balancing
load at the DIE level, the compute die sched group would be
regarded as group_has_spare.
In such case, inside find_busiest_group() the condition below
would be met:

        if (sds.prefer_sibling && local->group_type == group_has_spare &&
            busiest->sum_nr_running > local->sum_nr_running + 1)

and a balance will be triggered to pull task from Compute die
to Soc die, which is against the SOC atom strategy.

Consider this, remove the SD_PREFER_SIBLING flag of the MC domain,
so in Die domain the tasks would not be spread among die sched
groups.

Suggested-by: Ricardo Neri <ricardo.neri@intel.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>