mm/zswap: Add by2 compression/decompression
'by2' compression/decompression splits a swap page into 2 2k segments
and compresses those separately rather than one compression covering
the whole page.  The advantage of being able to do this is that each
2k chunk can be sent to separate hardware accelerators which can make
for significant improvements in tail latencies.

This allows the crypto compress/decompress calls to be polled for
completion using the acomp_alg and acomp_req poll() operations rather
than the standard waiting using acomp_ctx->wait.

This provides the opportunity to makes full use of hardware
compress/decompress engines without requiring the use of interrupts
and waiting, which add too much overhead to the pipeline; typically
these hw operations complete immediately and in the vast majority of
cases, a single call to crypto_acomp_poll() returns a completed
operation - adding the overhead of interrupt handling and waiting for
completions simply adds unnecessary delays and cancels the gains of
using the hw acceleration.

To set this up, use the by2_enabled zswap module param:

  echo 1 > /sys/module/zswap/parameters/by2_enabled

This is typically used with async mode in the driver.  For example
using IAA:

  echo async > /sys/module/iaa_crypto/parameters/iaa_sync_mode

mm/zswap: Add 'selective' by2 compression/decompression

'selective' by2 compression mode applies by2 compression only to those
pages larger than a given threshold.  The page is first compressed,
and the compressed size is compared against a threshold - if the size
is larger than the threshold, the compression is discarded and by2
compressed instead.

The cost of the extra compression in the case where the by2
compression is performed is assumed to be offset by the gain in
decompression efficiency when the page is decompressed.

A by2_threshold module param is used for this purpose.  To set the
by2_threshold, write the desired threshold to the by2_threshold file:

  echo 2048 > /sys/module/zswap/parameters/by2_threshold

The default by2_threshold is 0, meaning that when by2_enabled is set,
all pages will by split using by2.

If the by2_treshold is set to 4096, no pages are split using by2, but
the async poll() interface is still used.

byN, cleanup by2 and make byN

Also remove byN_enabled and replace by byN

byN 0 = disabled

refactor using do_byN

change cond_resched() to cpu_relax()

We're not coordinating with e.g. a wq and will get hw timeout if
things go too long, and aren't preempt-disabled so we can be
preempted, so don't need to coordinate with scheduler.

remove byN memsets for compress/decompress

The init_sg* functions already do this, an we set req.

fix bug missing last byte of odd byN

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>