perf/x86/intel: Add dedicated enable/disable event
The enable/disable event functions are shared among different X86
families, but the way to enable/disable an event may be different among
different X86 families. For example, AMD supports counter pair, but
Intel doesn't have the feature. To handle the differentiation, a check
has to be added in the common code. With more and more unique features
added, the code will become ugly and low efficient.

Use dedicated enable/disable event functions to replace the generic one
for Intel platforms. Drop the counter pair support, which is not
available for the Intel platforms.

No functional changes.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>