tty: Prevent writing chars during tcsetattr TCSADRAIN/FLUSH
If userspace races tcsetattr() with a write, the drained/flushed
condition might not be guaranteed. As ongoing transmission might
produce garbage during HW's ->set_termios() call, writes of additional
characters must be prevented from reaching the lower layer. Take tty's
write lock to achieve it while waiting for the lower layer to sent the
pending characters out.

The wait for all-writers-gone must be placed outside of tty's write
lock to avoid deadlock which makes it impossible to use
tty_wait_until_sent(). Retry the write lock if a racing write is
detected.

Signed-off-by: Ilpo Järvinen <ilpo.jarvinen@linux.intel.com>