platform/x86/intel/ifs: Add Array Scan support to IFS
Array BIST test will be first supported in the Emerald Rapids(EMR)
platform. Array BIST test searches for faults in the array portion of
silicon. The term array in the context of IFS could be caches,
register files or any repeated on-core structures of the CPU.

Array BIST checks portions of the core which are not tested by Scan test.
Array test uses internal circuits to run test  and no external tests
files are required.

Originally-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Jithu Joseph <jithu.joseph@intel.com>