ACPI: PNP: Do not create pnp device if its _HID is not PNP0C02
Currently the pfr_update and pfr_telemetry drivers could not
be loaded on SPR/GNR platforms. The cause of this issue is that,
there are no platform devices created for PFRU ACPI object INTC1080
and its corresponding telemetry object INTC1081. The reason why the
platform devices are not created is because if these objects have
_CID of PNP0C02, the acpi_pnp driver will take care of them, and
create pnp devices rather than platform devices.

To fix this issue, enhance the restriction that, if these ACPI
objects do not have _HID of PNP0C02, even if they have _CID of
PNP0C02, do not create the pnp devices for them. In this way, their
platform devices could be created, and PFRU driver could be loaded
successfully.

This is a trial patch to workaround this issue. The final upstream
version might be different from this.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>