x86/microcode/intel: Prepare save_microcode() to allow saving scanned image
Currently kernel saves the microcode for early loading support. The save
should be done only if the apply microcode was successful, if apply
microcode fails, keep the previous image as current image.

Save microcode must be split into two steps.

1. Save the scanned image in a separate variable after a ops->request_microcode_fw()
2. After apply microcode is successful, move it to current image referred
   by intel_ucode_patch.

If apply fails, just keep the earlier image as current image. This is the
image that will be used during an early load due to resume.

Change save_microcode() to take a parameter, so it can either save the
scanned_image, or update the current image pointed by intel_ucode_patch.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>