x86/microcode: Add support for uniform microcode loading
New Intel CPUs have a capability to update microcode across the whole
system. This eliminates a lot of the races with taking NMI's during the
update.

Add generic support to extend this mechanism across vendors. If vendor
supports this capability, then the update will be performed only to the
scope of CPUs.

The scope of update can be one of the following.

Platform Scope	- Updating one CPU will update all CPUs in the platform.
Package Scope	- Updating one CPU in the package will update all cores
		  within the package.
Core Scope 	- This is the legacy behavior, i.e every core needs an
		  update.

Extend late loading microcode to only run the update on one of the CPUs in
the scope.