x86/microcode: Don't use NMI rendezvous when UNIFORM update is supported
When processor reports supporting UNIFORM microcode update, there is no
reason to do perform the NMI late dance. Uniform update provides
architectural gaurantees that the update will not be triggered in the
middle of an instruction execution.

Set a flag to indicate to indicate that NMI late dance is required. When
UNIFORM loading is supported, clear this to avoid the NMI flows.

TBD: Now that we have two capabilities that the vendor can expose,
safe_late_loading and NMI rendezvous is required, its probably better to
expose this as some capabitlity bit flags. This is being deferred for a
later cleanup.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>