x86/microcode/intel: Add a generic pre-update vendor callback
When performing a new microcode reload via legacy interface, switch
the config to perform auto commit. This simplifies the update mecanism to
not perform another commit update on every logical core.

Add a new generic interface that vendor can support, so we can easily plug
these operations in. Later when we support microcode staging, the same
pre-update can be used to stage microcode as well.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>