x86/microcode/intel: Handle legacy update when deferred commit is enabled
When deferred commit is enabled, after a wrmsr 0x79 is complete, SW needs
to perform a manual commit. Since we aren't changing legacy behavior of the
sysfs reload file, comprehend this change.

During init, make sure the config is auto-commit, but don't lock the
interface. In subsequent patches when no commit update happens, we can
switch to manual commit on demand.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>