platform/x86/intel/ifs: Optimize Hash loading
In Scenarios where reduced memory support is supported by hardware. Only
strides containing distinct versions in header need to load hashes.

If the currently loaded version matches that of the file being loaded,
proceed directly to authenticate and copy chunks.

For gen1 (GNR A0) authenticate and copy can also be skipped if the
version matches as there is no concept of stride.

Signed-off-by: Jithu Joseph <jithu.joseph@intel.com>