perf/x86/intel: Support arch v6 part 2 (POC)
Support
- RDPMC_USER_DISABLE

Only disable the USR_RDPMC when the counter schedule in to prevent the
leakage of the system-wide counters.

Nothing changes in the counter schedule out, because
- The RDPMC user may disable the event first, then read the event.
  If we unconditionally disable USR_RDPMC in schedule out, noting can be
  read.

This patch still rely on the ("perf: Reset the dirty counter to prevent
the leak for an RDPMC task") for the per-thread counters.

Todo:
- New MSR location

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>