x86/asm: Add RDMSRLIST/WRMSRLIST insns support
Add two primitives to support the new RDMSRLIST/WRMSRLIST instructions.

The RDMSRLIST and WRMSRLIST are instructions to read/write a
software-provided list of up to 64 MSRs to/from memory in a performance
manner.

Introduce a new dedicated file asm/msr-list.h rather than asm/msr.h.
The msr.h is included everywhere. It’s not necessary to include the
MSRLIST instructions which is only available for newer platforms.

The primitives will be used by the later patches.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>