x86/msrlist: Add helper functions
To facilitate the use of the MSRLIST instructions, two helper functions
and a global per-cpu struct are added and will be used in the later
patch.

The global struct is to store the number of MSRs and their address and
values. It can help to simplify the design and get better performance.
The user doesn't need to allocate the space at runtime for the MSRLIST
instructions. It has to be per-CPU, since MSRs can be CPU scope.
The MSRLIST instructions are used in the critical path. The preemption
have been disabled. There should be no competition for the shared
variables. Extra lock is not required when accessing the global
variables.

The wrmsrl_batch() can be used to replace the wrmsrl() function. For a
platform which supports the WRMSRLIST instruction, the MSR address and
its value are stored in the global per-cpu struct. They will be written
by the MSRLIST instruction together in the new msrlist_writel() function
later. For the platform which doesn't support the WRMSRLIST instruction,
fall back to the wrmsrl(). There is nothing changed.

The wrmsrl() has slightly better performance than the WRMSRLIST
instruction when the number of MSRs is less than 3. Fall back to the
wrmsrl() for size 1 and 2.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>