KVM: TDX: Invoke machine check handler in #MSMI case
If there's an #MSMI (Machine Check System Management Interrupt) with
exit_qualification bit 0 set, this #MSMI is morphed by an #MC in TD
guest. Distinguish it with other EXIT_REASON_OTHER_SMI and only handle
MSMI as a user #MC.

This can also help distinguish with the TDX_SEAMCALL_UD check in the
coming commit.

Signed-off-by: Chenyi Qiang <chenyi.qiang@intel.com>