x86/sgx: Add SGX2 support preventing new initializations of EPC pages
Introduce mechanism to prevent new initializations of EPC pages for
SGX2.

This patch is SGX2 seamless part picked from:
https://lore.kernel.org/linux-sgx/20220315010300.10199-2-cathy.zhang@intel.com/

Covered by "[RFC PATCH v2 01/10] x86/sgx:Introduce mechanism to prevent
new initializations of EPC pages"

Use a flag to indicate when SGX EPC pages are "locked", which means
it's not allowed to allocate new EPC page for use. Check it in all
paths that can initialize an EPC page. Use SRCU to ensure that the
flag is visible across the system before proceeding with an update.

Add checks to all sites that call SGX2 instructions that can transition
pages from unused to initialized to ensure that the SRCU lock is held.

Co-developed-by: Cathy Zhang <cathy.zhang@intel.com>
Signed-off-by: Cathy Zhang <cathy.zhang@intel.com>
Signed-off-by: Qian Wen <qian.wen@linux.intel.com>