tools/power intel-speed-select: remove bogus check
struct isst_id *id is a pointer, comparing it with Zero is wrong.

The check there is to make sure the id->pkg/id->die is set to -1 when it
is illegal/unavailable. And comparing with
MAX_PACKAGE_COUNT/MAX_DIE_PER_PACKAGE is sufficient.

Thus remove the bogus check.

This is a standalone patch because the original commit is in upstream.
This fixes issue #1 and #2 in http://power-sh.sh.intel.com/power/rui-debug/coverity/cov-isst-html-orig/

Signed-off-by: Zhang Rui <rui.zhang@intel.com>