KVM: TDX: don't call TDH_VP_CREATE if TD hasn't been initialized
The vcpu initialization (e.g. TDH_VP_CREATE and TDH_VP_ADDCX seamcalls)
requires that TD has been initialized. Add a check of before proceeding
to making those seamcalls to initialize vcpu in the TDX module.

One example is that during live migraiton, the destination TD gets
initialized after the TD-scope state is imported, which is at a later
stage. So the vCPU initialization is delayed to that later point.

The allocation of tdvpr and tdvpx pages can be done earlier in
tdx_td_vcpu_init, so that they can be directly used later on the
performance critical path in live migration.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>