perf/x86/intel: Add TSX remote killer support (POC)
Hidden conflicts is one of the top performance limiters in TSX programs.

For stores it’s not easy to find the TSX remote killer. The trick is to
correlate the data address of two events to find the killer (sampling
the STORE_TSX_KILL event and MEM_INST_RETIRED.ALL_STORES).

For the STORE_TSX_KILL event, the PA of the killer is stored in the
MSR_TSX_CONFLICT_ADDR. This patch saves it into the phys_addr and
exposes it to user space.

TODO: For the STORE_TSX_KILL event, the addr is from the DLA of itself.
But the phys_addr is from MSR_TSX_CONFLICT_ADDR.
Need to check whether we should change the addr to killer's VA.
If the event's DLA/PA is also important, may have to introduce a new
sample type.

Notes:
- No MSR address for the MSR_TSX_CONFLICT_ADDR yet
- No event coding for the STORE_TSX_KILL event yet (Add it in the event
  constraint table.)
- No data src encoding (for local and remote (later generation))

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>